+++ backend/user/src/main/java/com/fengf/demo/springcloud/user/pojo/User.java
@@ -16,7 +16,7 @@ import javax.validation.constraints.NotNull;
 @ApiModel("用户基本信息")
 public class User {
     @ApiModelProperty("用户id")
-    private Integer userId;  // 用户id
+    private Long userId;  // 用户id
 
     @NotNull(message = "传入的userName为null，请传值")
     @NotEmpty(message = "传入的userName为空字符串，请传值")
